+++ src/ai/AiGlobalConfig.ts
@@ -1,5 +1,6 @@
 import {AiModelConfig} from "./core/AiModelConfig.ts";
 import {AiModelFactory} from "./AiModelFactory.ts";
+import {AIBubbleMenuItem} from "../components/bubbles/types.ts";
 
 export interface AiMenu {
     icon: string,
@@ -15,9 +16,10 @@ export interface AiGlobalConfig {
     models: Record<string, AiModelConfig>,
     modelFactory?: AiModelFactory,
     onTokenConsume?: (modelName: string, modelConfig: AiModelConfig, count: number) => void,
-    onCreateClientUrl?: (modelName: string, modelConfig: AiModelConfig, onFinished: (url: string) => void) => void,
+    onCreateClientUrl?: (modelName: string, modelConfig: AiModelConfig, onSuccess: (url: string) => void, onFailure: () => void) => void
     bubblePanelEnable?: boolean,
     bubblePanelModel?: string,
+    bubblePanelMenus?: AIBubbleMenuItem[],
     menus?: AiMenu[],
     commands?: AiMenu[],
     codeBlock?: {