+++ src/ai/core/AiModel.ts
@@ -1,40 +1,57 @@
 import {AiMessageListener} from "./AiMessageListener.ts";
 import {AiClient} from "./AiClient.ts";
-import {Editor} from "@tiptap/core";
 import {AiGlobalConfig} from "../AiGlobalConfig.ts";
 import {AiModelConfig} from "./AiModelConfig.ts";
-
-export interface ChatController {
-    stop: () => void
-}
+import {InnerEditor} from "../../core/AiEditor.ts";
 
 
 export abstract class AiModel {
 
-    editor: Editor;
-    globalConfig: AiGlobalConfig;
-    aiModelName: string;
-    aiModelConfig: AiModelConfig;
+    public editor: InnerEditor;
+    public globalConfig: AiGlobalConfig;
+    public aiModelName: string;
+    public aiModelConfig: AiModelConfig;
 
-
-    constructor(editor: Editor, globalConfig: AiGlobalConfig, aiModelName: string) {
+    protected constructor(editor: InnerEditor, globalConfig: AiGlobalConfig, aiModelName: string) {
         this.editor = editor;
         this.globalConfig = globalConfig;
         this.aiModelName = aiModelName;
         this.aiModelConfig = globalConfig.models[aiModelName];
     }
 
-    chat(selectedText: string, prompt: string, listener: AiMessageListener): void {
-        const startFunc = (url: string) => {
+    chatWithPayload(payload: any, listener: AiMessageListener): void {
+        const onSuccess = (url: string) => {
             const aiClient = this.createAiClient(url, listener);
-            const promptMessage = this.wrapMessage(`${selectedText}\n${prompt}`);
-            aiClient.start(promptMessage)
+            aiClient.start(typeof payload === "string" ? payload : JSON.stringify(payload))
+        }
+        const onFailure = () => {
+            listener?.onStop();
         }
 
         if (this.globalConfig.onCreateClientUrl) {
-            this.globalConfig.onCreateClientUrl(this.aiModelName, this.aiModelConfig, startFunc)
+            this.globalConfig.onCreateClientUrl(this.aiModelName, this.aiModelConfig, onSuccess, onFailure)
         } else {
-            startFunc(this.createAiClientUrl())
+            onSuccess(this.createAiClientUrl())
+        }
+    }
+
+
+    chat(selectedText: string, prompt: string, listener: AiMessageListener): void {
+        const onSuccess = (url: string) => {
+            const aiClient = this.createAiClient(url, listener);
+            const finalPrompt = prompt.includes("{content}") ? prompt.split('{content}').join(selectedText) : `${selectedText}\n${prompt}`
+            const payload = this.wrapPayload(finalPrompt);
+            aiClient.start(typeof payload === "string" ? payload : JSON.stringify(payload))
+        }
+
+        const onFailure = () => {
+            listener?.onStop();
+        }
+
+        if (this.globalConfig.onCreateClientUrl) {
+            this.globalConfig.onCreateClientUrl(this.aiModelName, this.aiModelConfig, onSuccess, onFailure)
+        } else {
+            onSuccess(this.createAiClientUrl())
         }
     }
 
@@ -51,13 +68,9 @@ export abstract class AiModel {
 
     /**
      * 封装消息，把 prompt 转换为协议需要的格式
-     * @param promptMessage
+     * @param prompt
      */
-    abstract wrapMessage(promptMessage: string): any;
+    abstract wrapPayload(prompt: string): any;
 
-
-    // call
-
-    // embeddings
 
 }