+++ src/components/AbstractBubbleMenu.ts
@@ -1,17 +1,16 @@
-import {AiEditorOptions, AiEditorEvent} from "../core/AiEditor.ts";
+import {AiEditorOptions, AiEditorEvent, InnerEditor} from "../core/AiEditor.ts";
 import {Editor, EditorEvents} from "@tiptap/core";
-import tippy from "tippy.js";
+import tippy, {Instance} from "tippy.js";
+import {BubbleMenuItem} from "./bubbles/types.ts";
+import {MenuRecord} from "./bubbles/items/MenuRecord.ts";
+import {t} from "i18next";
 
-export type BubbleMenuItem = {
-    id: string,
-    title?: string,
-    content: string,
-}
 
 export abstract class AbstractBubbleMenu extends HTMLElement implements AiEditorEvent {
 
     editor?: Editor;
     items: BubbleMenuItem[] = [];
+    tippyInstance?: Instance;
 
     protected constructor() {
         super();
@@ -22,21 +21,24 @@ export abstract class AbstractBubbleMenu extends HTMLElement implements AiEditor
     }
 
     connectedCallback() {
-
         this.innerHTML = `
             <div class="aie-bubble-menu">
                ${this.items!.map((item) => {
-            return `<div class="aie-bubble-menu-item ${this.isActive(item.id) ? 'active' : ''}" id="${item.id}">${item.content}</div>`
+            return `<div class="aie-bubble-menu-item ${this.isActive(item.id) ? 'active' : ''}" id="${item.id}">${item.icon}</div>`
         }).join('')}
             </div>
         `;
+
+        for (let item of this.items) {
+            item.holder = item.onInit?.((this.editor as InnerEditor).aiEditor, this.tippyInstance!, this);
+        }
 
         this.querySelector("div")!.addEventListener("click", (e) => {
             this.items.forEach((item) => {
                 const target = (e.target as any).closest(`#${item.id}`);
                 if (target) {
                     target.classList.contains("active") ? target.classList.remove("active") : target.classList.add("active")
-                    this.onItemClick(item.id);
+                    this.onItemClick(item);
                 }
             })
         })
@@ -46,7 +48,7 @@ export abstract class AbstractBubbleMenu extends HTMLElement implements AiEditor
             if (title) {
                 tippy(el, {
                     appendTo: this.closest(".aie-container")!,
-                    content: title,
+                    content: t(title),
                     theme: 'aietip',
                     arrow: true,
                     // trigger:"click",
@@ -56,13 +58,31 @@ export abstract class AbstractBubbleMenu extends HTMLElement implements AiEditor
         });
     }
 
-    onCreate(props: EditorEvents['create'], _: AiEditorOptions) {
-        this.editor = props.editor
+    set instance(value: Instance) {
+        this.tippyInstance = value;
     }
 
-    abstract onItemClick(id: string): void;
+    protected initItemsByOptions(allMenuItems: MenuRecord, optionItems?: (string | BubbleMenuItem)[]) {
+        if (optionItems && optionItems.length > 0) {
+            for (let key of optionItems) {
+                const linkMenuItem = typeof key === "string" ? allMenuItems.getItem(key) : key;
+                if (linkMenuItem) this.items.push(linkMenuItem);
+            }
+        } else {
+            this.items = allMenuItems.getAllItem()
+        }
+    }
 
-    abstract onTransaction(props: EditorEvents['transaction']): void
+    onCreate(createEvent: EditorEvents['create'], _: AiEditorOptions) {
+        this.editor = createEvent.editor
+    }
+
+    onItemClick(item: BubbleMenuItem) {
+        item.onClick?.((this.editor as InnerEditor).aiEditor, this.tippyInstance!, this, item.holder);
+    }
+
+    onTransaction(_transEvent: EditorEvents['transaction']) {
+    }
 
 }