+++ src/i18n/en.ts
@@ -3,6 +3,8 @@ export const en =  {
     "default":"Default",
     "delete": "Delete",
     "confirm": "Confirm",
+    "cancel": "Cancel",
+
 
     //menu tips
     "undo": "Undo",
@@ -18,6 +20,7 @@ export const en =  {
     "underline": "Underline",
     "strike": "Strikethrough",
     "link": "Link",
+    "source-code": "Source code",
     "code": "Inline code",
     "subscript": "Subscript",
     "superscript": "Superscript",
@@ -92,4 +95,31 @@ export const en =  {
     ////font
     "default-font-size":"Size",
     "default-font-family":"Default font",
+
+    ////code block
+    "codeblock-ai-auto-comment":"auto comment",
+    "codeblock-ai-code-explain":"code explain",
+
+
+    //// ai
+    "placeholder-tell-ai-what-to-do-next":"Tell AI what to do next? e.g., help me translate into English",
+    "tip-you-can-enter-text-above-or-select-the-operation-below":"Tip: You can enter text above or select the operation below",
+    "you-can-do-the-following":"You can do the following:",
+    "improve-writing":"Improve your writing",
+    "check-spelling-and-grammar":"Check spelling and grammar",
+    "make-shorter":"Make shorter",
+    "make-longer":"Make longer",
+    "translate":"Translate",
+    "summarize":"Summarize",
+    "ai-append":"Append",
+    "ai-replace":"Replace",
+    "ai-cancel":"Cancel",
+
+    ///ai menus
+    "ai-continuation":"Continuation",
+    "ai-optimization":"Optimization",
+    "ai-proofreading":"Proofreading",
+    "ai-translation":"Translation",
+
+
 }